Put all prompts under MIT license
fixed #15158
