+++ packages/ai-ide/src/common/orchestrator-chat-agent.ts
@@ -15,7 +15,6 @@
 // *****************************************************************************
 
 import { getJsonOfText, getTextOfResponse, LanguageModelRequirement, LanguageModelResponse } from '@theia/ai-core';
-import { PromptTemplate } from '@theia/ai-core/lib/common';
 import { inject, injectable } from '@theia/core/shared/inversify';
 import { ChatAgentService } from '@theia/ai-chat/lib/common/chat-agent-service';
 import { AbstractStreamParsingChatAgent, ChatSessionContext } from '@theia/ai-chat/lib/common/chat-agents';
@@ -23,42 +22,7 @@ import { MutableChatRequestModel, InformationalChatResponseContentImpl } from '@
 import { generateUuid, nls } from '@theia/core';
 import { ChatHistoryEntry } from '@theia/ai-chat/lib/common/chat-history-entry';
 
-export const orchestratorTemplate: PromptTemplate = {
-    id: 'orchestrator-system',
-    template: `{{!-- Made improvements or adaptations to this prompt template? We’d love for you to share it with the community! Contribute back here:
-https://github.com/eclipse-theia/theia/discussions/new?category=prompt-template-contribution --}}
-# Instructions
-
-Your task is to identify which Chat Agent(s) should best reply a given user's message.
-You consider all messages of the conversation to ensure consistency and avoid agent switches without a clear context change.
-You should select the best Chat Agent based on the name and description of the agents, matching them to the user message.
-
-## Constraints
-
-Your response must be a JSON array containing the id(s) of the selected Chat Agent(s).
-
-* Do not use ids that are not provided in the list below.
-* Do not include any additional information, explanations, or questions for the user.
-* If there is no suitable choice, pick \`Universal\`.
-* If there are multiple good choices, return all of them.
-
-Unless there is a more specific agent available, select \`Universal\`, especially for general programming-related questions.
-You must only use the \`id\` attribute of the agent, never the name.
-
-### Example Results
-
-\`\`\`json
-["Universal"]
-\`\`\`
-
-\`\`\`json
-["AnotherChatAgent", "Universal"]
-\`\`\`
-
-## List of Currently Available Chat Agents
-
-{{chatAgents}}
-`};
+import { orchestratorTemplate } from './orchestrator-prompt-template';
 
 export const OrchestratorChatAgentId = 'Orchestrator';
 const OrchestratorRequestIdKey = 'orchestatorRequestIdKey';
