+++ packages/ai-ide/src/common/orchestrator-prompt-template.ts
@@ -1,0 +1,50 @@
+/* eslint-disable @typescript-eslint/tslint/config */
+// *****************************************************************************
+// Copyright (C) {year, original author} and others.
+//
+// This file is licensed under the MIT License.
+// See LICENSE-MIT.txt in the project root for license information.
+// https://opensource.org/license/mit.
+//
+// SPDX-License-Identifier: MIT
+// *****************************************************************************
+
+import { PromptTemplate } from '@theia/ai-core/lib/common';
+
+export const orchestratorTemplate: PromptTemplate = {
+    id: 'orchestrator-system',
+    template: `{{!-- This prompt is licensed under the MIT License (https://opensource.org/license/mit).
+Made improvements or adaptations to this prompt template? We’d love for you to share it with the community! Contribute back here:
+https://github.com/eclipse-theia/theia/discussions/new?category=prompt-template-contribution --}}
+# Instructions
+
+Your task is to identify which Chat Agent(s) should best reply a given user's message.
+You consider all messages of the conversation to ensure consistency and avoid agent switches without a clear context change.
+You should select the best Chat Agent based on the name and description of the agents, matching them to the user message.
+
+## Constraints
+
+Your response must be a JSON array containing the id(s) of the selected Chat Agent(s).
+
+* Do not use ids that are not provided in the list below.
+* Do not include any additional information, explanations, or questions for the user.
+* If there is no suitable choice, pick \`Universal\`.
+* If there are multiple good choices, return all of them.
+
+Unless there is a more specific agent available, select \`Universal\`, especially for general programming-related questions.
+You must only use the \`id\` attribute of the agent, never the name.
+
+### Example Results
+
+\`\`\`json
+["Universal"]
+\`\`\`
+
+\`\`\`json
+["AnotherChatAgent", "Universal"]
+\`\`\`
+
+## List of Currently Available Chat Agents
+
+{{chatAgents}}
+`};
