+++ packages/ai-ide/src/common/universal-prompt-template.ts
@@ -1,17 +1,12 @@
+/* eslint-disable @typescript-eslint/tslint/config */
 // *****************************************************************************
-// Copyright (C) 2025 EclipseSource GmbH.
+// Copyright (C) {year, original author} and others.
 //
-// This program and the accompanying materials are made available under the
-// terms of the Eclipse Public License v. 2.0 which is available at
-// http://www.eclipse.org/legal/epl-2.0.
+// This file is licensed under the MIT License.
+// See LICENSE-MIT.txt in the project root for license information.
+// https://opensource.org/license/mit.
 //
-// This Source Code may also be made available under the following Secondary
-// Licenses when the conditions for such availability set forth in the Eclipse
-// Public License v. 2.0 are satisfied: GNU General Public License, version 2
-// with the GNU Classpath Exception which is available at
-// https://www.gnu.org/software/classpath/license.html.
-//
-// SPDX-License-Identifier: EPL-2.0 OR GPL-2.0-only WITH Classpath-exception-2.0
+// SPDX-License-Identifier: MIT
 // *****************************************************************************
 
 import { PromptTemplate } from '@theia/ai-core/lib/common';
@@ -19,7 +14,8 @@ import { CHAT_CONTEXT_DETAILS_VARIABLE_ID } from '@theia/ai-chat';
 
 export const universalTemplate: PromptTemplate = {
    id: 'universal-system',
-   template: `{{!-- Made improvements or adaptations to this prompt template? We’d love for you to share it with the community! Contribute back here:
+   template: `{{!-- This prompt is licensed under the MIT License (https://opensource.org/license/mit).
+Made improvements or adaptations to this prompt template? We’d love for you to share it with the community! Contribute back here:
 https://github.com/eclipse-theia/theia/discussions/new?category=prompt-template-contribution --}}
 
 You are an assistant integrated into Theia IDE, designed to assist software developers.
